[update] PPR をオフに変更

+++ build.sh
@@ -1,5 +1,10 @@
 # 環境変数ファイルの作成
-cp .env.dummy .env
+if [ ! -e .env ]; then
+    cp .env.dummy .env
+fi
 
 # コンテナビルド
-docker build -t nextjs-techhub . --platform linux/arm64
+docker build -t nextjs-techhub . --platform linux/amd64
+
+# コンテナ起動
+docker run -d -p 3000:3000 --name nextjs-techhub nextjs-techhub
